+++ goflow/engine.go
@@ -10,28 +10,26 @@ import (
 	"github.com/nyaruka/goflow/flows"
 	"github.com/nyaruka/goflow/flows/engine"
 	"github.com/nyaruka/goflow/services/webhooks"
+	"github.com/nyaruka/goflow/utils/httpx"
 	"github.com/nyaruka/mailroom/config"
 
 	"github.com/shopspring/decimal"
 )
 
-var httpClient *http.Client
 var eng, simulator flows.Engine
-var engInit, simulatorInit sync.Once
+var engInit, simulatorInit, webhooksHTTPInit sync.Once
+
+var webhooksHTTPClient *http.Client
+var webhooksHTTPRetries *httpx.RetryConfig
+
+var emailFactory engine.EmailServiceFactory
 var classificationFactory engine.ClassificationServiceFactory
 var airtimeFactory engine.AirtimeServiceFactory
 
-func init() {
-	// customize the default golang transport
-	t := http.DefaultTransport.(*http.Transport).Clone()
-	t.MaxIdleConns = 32
-	t.MaxIdleConnsPerHost = 8
-	t.IdleConnTimeout = 30 * time.Second
-	t.TLSClientConfig = &tls.Config{
-		Renegotiation: tls.RenegotiateOnceAsClient, // support single TLS renegotiation
-	}
-
-	httpClient = &http.Client{Transport: t, Timeout: time.Duration(15 * time.Second)}
+// RegisterEmailServiceFactory can be used by outside callers to register a email factory
+// for use by the engine
+func RegisterEmailServiceFactory(factory engine.EmailServiceFactory) {
+	emailFactory = factory
 }
 
 // RegisterClassificationServiceFactory can be used by outside callers to register a classification factory
@@ -54,8 +52,11 @@ func Engine() flows.Engine {
 			"X-Mailroom-Mode": "normal",
 		}
 
+		httpClient, httpRetries := webhooksHTTP()
+
 		eng = engine.NewBuilder().
-			WithWebhookServiceFactory(webhooks.NewServiceFactory(httpClient, webhookHeaders, 10000)).
+			WithWebhookServiceFactory(webhooks.NewServiceFactory(httpClient, httpRetries, webhookHeaders, config.Mailroom.WebhooksMaxBodyBytes)).
+			WithEmailServiceFactory(emailFactory).
 			WithClassificationServiceFactory(classificationFactory).
 			WithAirtimeServiceFactory(airtimeFactory).
 			WithMaxStepsPerSprint(config.Mailroom.MaxStepsPerSprint).
@@ -73,15 +74,53 @@ func Simulator() flows.Engine {
 			"X-Mailroom-Mode": "simulation",
 		}
 
+		httpClient, _ := webhooksHTTP() // don't do retries in simulator
+
 		simulator = engine.NewBuilder().
-			WithWebhookServiceFactory(webhooks.NewServiceFactory(httpClient, webhookHeaders, 10000)).
+			WithWebhookServiceFactory(webhooks.NewServiceFactory(httpClient, nil, webhookHeaders, config.Mailroom.WebhooksMaxBodyBytes)).
 			WithClassificationServiceFactory(classificationFactory).   // simulated sessions do real classification
-			WithAirtimeServiceFactory(simulatorAirtimeServiceFactory). // but faked airtime transfers
+			WithEmailServiceFactory(simulatorEmailServiceFactory).     // but faked emails
+			WithAirtimeServiceFactory(simulatorAirtimeServiceFactory). // and faked airtime transfers
 			WithMaxStepsPerSprint(config.Mailroom.MaxStepsPerSprint).
 			Build()
 	})
 
 	return simulator
+}
+
+func webhooksHTTP() (*http.Client, *httpx.RetryConfig) {
+	webhooksHTTPInit.Do(func() {
+		// customize the default golang transport
+		t := http.DefaultTransport.(*http.Transport).Clone()
+		t.MaxIdleConns = 32
+		t.MaxIdleConnsPerHost = 8
+		t.IdleConnTimeout = 30 * time.Second
+		t.TLSClientConfig = &tls.Config{
+			Renegotiation: tls.RenegotiateOnceAsClient, // support single TLS renegotiation
+		}
+
+		webhooksHTTPClient = &http.Client{
+			Transport: t,
+			Timeout:   time.Duration(config.Mailroom.WebhooksTimeout) * time.Millisecond,
+		}
+
+		webhooksHTTPRetries = httpx.NewExponentialRetries(
+			time.Duration(config.Mailroom.WebhooksInitialBackoff)*time.Millisecond,
+			config.Mailroom.WebhooksMaxRetries,
+			config.Mailroom.WebhooksBackoffJitter,
+		)
+	})
+	return webhooksHTTPClient, webhooksHTTPRetries
+}
+
+func simulatorEmailServiceFactory(session flows.Session) (flows.EmailService, error) {
+	return &simulatorEmailService{}, nil
+}
+
+type simulatorEmailService struct{}
+
+func (s *simulatorEmailService) Send(session flows.Session, addresses []string, subject, body string) error {
+	return nil
 }
 
 func simulatorAirtimeServiceFactory(session flows.Session) (flows.AirtimeService, error) {