+++ models/env.go
@@ -9,6 +9,8 @@ import (
 	"github.com/nyaruka/goflow/envs"
 	"github.com/nyaruka/goflow/flows"
 	"github.com/nyaruka/goflow/services/airtime/dtone"
+	"github.com/nyaruka/goflow/services/email/smtp"
+	"github.com/nyaruka/goflow/utils/httpx"
 	"github.com/nyaruka/mailroom/config"
 	"github.com/nyaruka/mailroom/goflow"
 
@@ -20,11 +22,18 @@ import (
 // Register a airtime service factory with the engine
 func init() {
 	// give airtime transfers an extra long timeout
-	httpClient := &http.Client{Timeout: time.Duration(120 * time.Second)}
+	airtimeHTTPClient := &http.Client{Timeout: time.Duration(120 * time.Second)}
+	airtimeHTTPRetries := httpx.NewFixedRetries(5, 10)
+
+	goflow.RegisterEmailServiceFactory(
+		func(session flows.Session) (flows.EmailService, error) {
+			return orgFromSession(session).EmailService(http.DefaultClient)
+		},
+	)
 
 	goflow.RegisterAirtimeServiceFactory(
 		func(session flows.Session) (flows.AirtimeService, error) {
-			return orgFromSession(session).AirtimeService(httpClient)
+			return orgFromSession(session).AirtimeService(airtimeHTTPClient, airtimeHTTPRetries)
 		},
 	)
 }
@@ -34,6 +43,7 @@ type OrgID int
 const (
 	NilOrgID = OrgID(0)
 
+	configSMTPServer    = "smtp_server"
 	configDTOneLogin    = "TRANSFERTO_ACCOUNT_LOGIN"
 	configDTOneToken    = "TRANSFERTO_AIRTIME_API_TOKEN"
 	configDTOnecurrency = "TRANSFERTO_ACCOUNT_CURRENCY"
@@ -106,8 +116,18 @@ func (o *Org) ConfigValue(key string, def string) string {
 	return strVal
 }
 
+// EmailService returns the email service for this org
+func (o *Org) EmailService(httpClient *http.Client) (flows.EmailService, error) {
+	connectionURL := o.ConfigValue(configSMTPServer, config.Mailroom.SMTPServer)
+
+	if connectionURL == "" {
+		return nil, errors.New("missing SMTP configuration")
+	}
+	return smtp.NewServiceFromURL(connectionURL)
+}
+
 // AirtimeService returns the airtime service for this org if one is configured
-func (o *Org) AirtimeService(httpClient *http.Client) (flows.AirtimeService, error) {
+func (o *Org) AirtimeService(httpClient *http.Client, httpRetries *httpx.RetryConfig) (flows.AirtimeService, error) {
 	login := o.ConfigValue(configDTOneLogin, "")
 	token := o.ConfigValue(configDTOneToken, "")
 	currency := o.ConfigValue(configDTOnecurrency, "")
@@ -115,7 +135,7 @@ func (o *Org) AirtimeService(httpClient *http.Client) (flows.AirtimeService, err
 	if login == "" || token == "" {
 		return nil, errors.Errorf("missing %s or %s on DTOne configuration for org: %d", configDTOneLogin, configDTOneToken, o.ID())
 	}
-	return dtone.NewService(httpClient, login, token, currency), nil
+	return dtone.NewService(httpClient, httpRetries, login, token, currency), nil
 }
 
 // gets the underlying org for the given engine session