+++ models/msgs.go
@@ -320,13 +320,16 @@ func NewOutgoingMsg(orgID OrgID, channel *Channel, contactID ContactID, out *flo
 	}
 
 	// populate metadata if we have any
-	if len(out.QuickReplies()) > 0 || out.Templating() != nil {
+	if len(out.QuickReplies()) > 0 || out.Templating() != nil || out.Topic() != flows.NilMsgTopic {
 		metadata := make(map[string]interface{})
 		if len(out.QuickReplies()) > 0 {
 			metadata["quick_replies"] = out.QuickReplies()
 		}
 		if out.Templating() != nil {
 			metadata["templating"] = out.Templating()
+		}
+		if out.Topic() != flows.NilMsgTopic {
+			metadata["topic"] = string(out.Topic())
 		}
 		m.Metadata = null.NewMap(metadata)
 	}
@@ -862,6 +865,7 @@ func CreateBroadcastMessages(ctx context.Context, db Queryer, rp *redis.Pool, or
 			templateCtx := types.NewXObject(map[string]types.XValue{
 				"contact": flows.Context(org.Env(), contact),
 				"fields":  flows.Context(org.Env(), contact.Fields()),
+				"globals": flows.Context(org.Env(), sa.Globals()),
 				"urns":    flows.ContextFunc(org.Env(), contact.URNs().MapContext),
 			})
 			text, _ = excellent.EvaluateTemplate(org.Env(), templateCtx, template, nil)
@@ -873,7 +877,7 @@ func CreateBroadcastMessages(ctx context.Context, db Queryer, rp *redis.Pool, or
 		}
 
 		// create our outgoing message
-		out := flows.NewMsgOut(urn, channel.ChannelReference(), text, t.Attachments, t.QuickReplies, nil)
+		out := flows.NewMsgOut(urn, channel.ChannelReference(), text, t.Attachments, t.QuickReplies, nil, flows.NilMsgTopic)
 		msg, err := NewOutgoingMsg(org.OrgID(), channel, c.ID(), out, time.Now())
 		msg.SetBroadcastID(bcast.BroadcastID())
 		if err != nil {